+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added 
+- Set longer timeout to process the images [#182](https://github.com/rokwire/gateway-building-block/issues/182)
+
 ## [2.19.1] - 2026-01-28
 ### Fixed
 - Fix Webtools event image handling [#178](https://github.com/rokwire/gateway-building-block/issues/178)